Add files via upload
There one meeting room in a firm. There are N meetings in the form of (start[i], end[i]) where start[i] is start time of meeting, and end[i] is finish time of meeting i . what is the maximum number of meetings that can be accommodated in the meeting room when only one room at a particular time?